Updates from upstream, and use resource for image
There was a new upstream release for 0.2.1, including new images.

This also switches to using Juju resources for the JupyterHub image.
We may also want to switch to resources for the TensorFlow images but
they are quite large and maybe not ideal for going through the staging
server.  Also, since the form displays the entire image path, including
the repo, it may seem like we're deviating more from upstream, though we
could certainly change how they're presented in the form.